lab3/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.9
httpx==0.27.0
asyncpg==0.29.0
pytest==8.1.1
pytest-asyncio==0.23.6
testcontainers[postgres]==3.7.1
sqlalchemy==1.4.52
psycopg2-binary==2.9.9
